refactor: extract ingest orchestration
