+++ docs/agent-control/acceptance-matrix.md
@@ -19,7 +19,7 @@ and validation gates.
 | REQ-12 | Ship a native behavior layer so agents use Cortex automatically: plugin skills with trigger descriptions and a session bootstrap that survives new sessions, clear, and compaction. | Plugin directory has only the MCP manifest; Codex integration is a two-line AGENTS.md section; no skills, hooks, or marketplace entry. | `plugins/cortex` skills/hooks/manifests, root `.claude-plugin/marketplace.json`, `cortex init` AGENTS.md upgrade, `sync-release-version` coverage. | New root tests (session bootstrap, manifests, skill lint); init-agents tests; `release:sync-version --check`; full root + MCP suites. | Not ready |
 | REQ-13 | Keep Enterprise while enforcing a safe local trust boundary for remote organization content, license credentials, secrets, privileged control, and project-scoped daemon services. | Enterprise remains available behind explicit verified enrollment. Remote identifiers and persisted paths are untrusted, bearer endpoints require TLS/loopback, root control uses package-owned code, user-global artifacts/events are credential-bound, and project services are registered by normalized root. | WO-024 and context packet 014 across skill sync/rotation, license verification, enterprise setup/govern CLI, trusted packaging, daemon project/host services, and negative regression tests. | Security and release reviews: no blocker/major. MCP 413/413; root 254/254 plus context regressions 81/81; package 380 entries; version, diff, TypeScript, extracted-package smoke, and Cortex refresh gates. | Accepted locally |
 | REQ-14 | Ship the 2.4.1 security release without known npm dependency advisories. | WO-025 resolves the baseline one high frontend finding and 13 MCP vulnerable package nodes, including the critical tar chain and upstream no-automatic-fix Transformers dependencies. Release and Pages workflows now fail on any low-or-higher npm advisory. | Patched direct versions plus bounded transitive overrides where upstream ranges do not admit the fix; Node floor aligned to 20.9 LTS. | All committed lockfiles audit at zero; clean installs; frontend build; native embedding/image/ZIP smoke; MCP 413/413; root 254/254 plus 81/81 context regressions; version/workflow/package/diff gates. | Accepted locally; default-branch alerts close after merge |
-| REQ-15 | Keep Cortex maintainable by modularizing the CLI and ingest pipeline without changing public, security, data, determinism, or memory contracts. | WO-026 freezes the 2.4.1 contract. WO-027 replaces the 1,862-line CLI with a 34-line executable and 18 cohesive modules. WO-028 makes `scaffold/scripts/lib/ingest/` canonical, reduces both ingest entrypoints to thin executable wrappers, and extracts constants, arguments, paths/files, IO, chunks/modules, relations, projects, and incremental state without changing frozen output. | WO-029 through WO-031 and context packet 019: parser/worker/pipeline extraction, safe managed-file cleanup, and integrated patch-release validation. R16 records pre-existing ingest filesystem-containment hardening outside the behavior-preserving program. | WO-028: syntax 13/13; focused 61/61; root 81/81 + 300/300; MCP 413/413; unchanged four WO-026 digests plus sequential/parallel and worker-failure byte identity; 409-entry package with all 11 canonical modules and executable wrapper; Cortex update/doctor 8/8; five required reviewer roles closed. | WO-026 through WO-028 accepted locally; WO-029 ready; program in progress |
+| REQ-15 | Keep Cortex maintainable by modularizing the CLI and ingest pipeline without changing public, security, data, determinism, or memory contracts. | WO-026 freezes the 2.4.1 contract. WO-027 replaces the 1,862-line CLI with a 34-line executable and 18 cohesive modules. WO-028 makes `scaffold/scripts/lib/ingest/` canonical. WO-029 leaves a 76-line stage composition over 15 canonical modules, with parser registry/composition, worker streaming/protocol, and bounded state-passing stages extracted without changing frozen behavior. | WO-030 and WO-031 via packet 020: safe managed-file ownership/cleanup and integrated patch-release validation. R16 records pre-existing ingest filesystem-containment hardening outside this behavior-preserving program. | WO-029: syntax 19/19; focused 63/63; root 81/81 + 302/302; MCP 413/413; unchanged four digests plus sequential/parallel and worker-failure byte identity; all 17 traces; 413-entry package; three-run Cortex/Angular median RSS deltas +2.04%/-0.35%; five required reviewer roles closed after a verbose-window major was fixed. | WO-026 through WO-029 accepted locally; WO-030 ready; program in progress |
 
 ## Global Acceptance Requirements
