+++ docs/agent-control/agent-work-orders.md
@@ -33,8 +33,8 @@ fresh context packet before implementation starts.
 | WO-026 | Control Manager + Validation | Characterize CLI, Enterprise, ingest, packaging, and memory behavior before modularization. | CLI/Enterprise subprocess contracts, multilingual ingest fixture and normalized hashes, worker/trace coverage, package/install evidence, three-run memory medians, and focused packet 017. | Root 81/81 context regressions + 290/290 Node tests; MCP 413/413; package 380 entries and clean install/init; six benchmark samples `ok`; syntax/version/diff gates. | `v2.4.1` baseline; REQ-15; R14, R15 | Accepted locally |
 | WO-027 | CLI and Scaffold | Split `bin/cortex.mjs` into cohesive CLI modules while preserving command, stream, exit, JSON, compatibility-export, daemon, and Enterprise trust contracts. | Thin executable plus `bin/cli/` helpers, router, and command modules; focused packet `017-cli-modularization.md`. | Focused 58/58; root 81/81 + 293/293; MCP 413/413; 398-entry package with all 18 CLI modules; syntax/diff/Cortex gates; five-role review closure. | WO-026 | Accepted locally |
 | WO-028 | Parsers and Ingest | Establish the packaged scaffold as the single ingest source of truth and extract pure ingest responsibilities. | Thin root/scaffold entrypoints plus 11 canonical modules under `scaffold/scripts/lib/ingest/`; packet `018-canonical-ingest-pure-modules.md`; baseline `wo-028-canonical-ingest-baseline.md`. | Syntax 13/13; focused 61/61; root 81/81 + 300/300; MCP 413/413; 409-entry package with every canonical module; Cortex update/doctor 8/8; five-role review closure. | WO-026; WO-027 accepted locally | Accepted locally |
-| WO-029 | Parsers and Ingest | Extract ingest parser composition, worker scheduling/protocol, streaming result consumption, and pipeline orchestration without changing ordering, fallback, trace, or retention behavior. | Parser/worker/pipeline modules and bounded explicit stage composition; focused packet `019-ingest-orchestration-workers.md`. | Worker crash/timeout/fallback tests; output equivalence; trace compatibility; repeated memory evidence; full root suite. | WO-028 accepted locally | Ready |
-| WO-030 | CLI and Scaffold + Security and Privacy | Add a managed-scaffold manifest and safe obsolete-file cleanup for forced upgrades. | Versioned ownership manifest, contained cleanup logic, preservation and traversal/symlink tests. | Init/migration/Enterprise permission tests; stale-file bootstrap regression; package and clean-upgrade smokes; security review. | WO-027, WO-029 | Blocked |
+| WO-029 | Parsers and Ingest | Extract ingest parser composition, worker scheduling/protocol, streaming result consumption, and pipeline orchestration without changing ordering, fallback, trace, or retention behavior. | Fifteen canonical ingest modules, 76-line explicit stage composition, thin parser compatibility entry, baseline `wo-029-ingest-orchestration-baseline.md`, and packet `020-managed-scaffold-upgrade-hygiene.md`. | Syntax 19/19; focused 63/63; root 81/81 + 302/302; MCP 413/413; 413-entry package; all four digests/equivalence contracts; repeated memory 6/6 with medians inside 5%; five-role review closure. | WO-028 accepted locally | Accepted locally |
+| WO-030 | CLI and Scaffold + Security and Privacy | Add a managed-scaffold manifest and safe obsolete-file cleanup for forced upgrades. | Versioned ownership manifest, contained cleanup logic, preservation and traversal/symlink tests; packet `020-managed-scaffold-upgrade-hygiene.md`. | Init/migration/Enterprise permission tests; stale-file bootstrap regression; package and clean-upgrade smokes; security review. | WO-027, WO-029 accepted locally | Ready |
 | WO-031 | Validation + Release and Distribution | Run integrated modularization acceptance and prepare the patch release. | Full validation evidence, package install/bootstrap/search smoke, review closure, and `2.4.2` metadata only after acceptance. | Full root/MCP suites; audits; version sync; package extraction; doctor/search/update; memory comparison; all required reviewer sign-offs. | WO-027 through WO-030 | Blocked |
 
 ## Manager Assignment Checklist
