+++ docs/agent-control/handoff-ledger.md
@@ -16,6 +16,7 @@ next day rollover. This file keeps the template and open rows only.
 
 | Date | Agent/Team | Assignment | Context Packet | Owned Files/Entities | Output | Tests | Risks | Manager Decision |
 |---|---|---|---|---|---|---|---|---|
+| 2026-07-29 | Control Manager implementer; independent Code Quality/Integration, Contract/Validation, and Security/Privacy reviewers | WO-029 ingest orchestration extraction | `docs/agent-control/context-packets/019-ingest-orchestration-workers.md` | Parser registry/composition, worker streaming/protocol, pipeline stages/composition, ingest tests, WO-029 baseline, packet 020, control metadata | Extracted 15 canonical ingest modules behind a 76-line composition and one-line parser compatibility entry. Review found a major verbose overlap-window state omission; it was fixed with a subprocess regression. All five roles pass with no remaining blocker, major, or minor findings. | Syntax 19/19; focused 63/63; context regressions 81/81; root Node 302/302; MCP 413/413; unchanged four digests, sequential/parallel and worker-failure byte identity, 17 trace points; 413-entry package; memory 6/6 with Cortex/Angular median RSS deltas +2.04%/-0.35%; diff clean; Cortex update zero failures, indexed pattern evidence 7/7 with 11 configured-scope exclusions, doctor 8/8, watcher stopped | R14 mitigated locally through WO-029; R16 unchanged; R15 next | Accepted locally; WO-030 Ready |
 | 2026-07-29 | Control Manager implementer; independent Code Quality/Integration, Contract/Validation, and Security/Privacy reviewers | WO-028 canonical ingest source and pure modules | `docs/agent-control/context-packets/018-canonical-ingest-pure-modules.md` | Both ingest entrypoints, `scaffold/scripts/lib/ingest/*.mjs`, ingest/copy-fixture tests, WO-028 baseline, packet 019, control metadata | Established the packaged scaffold as the only implementation; reduced entrypoints to 14/16-line executable wrappers; extracted 11 canonical modules. Review found one minor stale line-count reference, fixed and re-reviewed. All five roles closed with no blocker/major findings. Security recorded unchanged source/output containment residuals as R16. | Syntax 13/13; focused 61/61; context regressions 81/81; root Node 300/300; MCP 413/413; unchanged full/changed hashes and worker/parallel byte identity; 409-entry package with all modules and executable wrapper; diff clean; Cortex update, doctor 8/8, watcher stopped; pattern evidence attempted with configured-scope exclusions documented | R14 pure/canonical portion mitigated; R16 accepted pre-existing | Accepted locally; WO-029 Ready |
 | 2026-07-29 | Control Manager implementer; independent Code Quality, Contract, Security/Privacy, Integration, Validation reviewers | WO-027 CLI modularization | `docs/agent-control/context-packets/017-cli-modularization.md` | `bin/cortex.mjs`, `bin/cli/*.mjs`, CLI/scaffold/Enterprise contract tests, plan/control metadata, packet 018 | Replaced the 1,862-line CLI with a 34-line executable, explicit router, and 18 cohesive modules. Project runtime and package-trusted Enterprise resolution are physically separate. Review iteration fixed blank-line secret input, split a secondary mixed-responsibility module, centralized project runtime loading, decoupled a source-spelling test, and added direct package-trusted install resolution coverage; every role re-reviewed clean. | Focused 58/58; root 81/81 + 293/293; MCP 413/413; all syntax checks; clean diff; 398-entry pack with every CLI module and executable mode; Cortex update/pattern evidence/doctor 8/8; watcher stopped | R14 CLI portion mitigated; ingest risk remains through WO-028/WO-029 | Accepted locally; WO-028 Ready |
 | 2026-07-28 | Control Manager implementer; independent Code Quality, Contract, Security/Privacy, Integration, Validation reviewers | WO-026 modularization characterization and baseline | `docs/agent-control/context-packets/016-cli-ingest-modularization.md` | CLI/Enterprise subprocess contracts, multilingual ingest fixture/hashes, worker fallback/trace tests, ignored memory evidence, package baseline, packet 017, control metadata | Froze the 2.4.1 contract without runtime refactoring. Review majors added successful handler/trusted-runtime/secret/config and 56-file worker-failure pipeline coverage; all findings were fixed and all five roles signed off. | Focused 46/46; root 81/81 + 290/290; MCP 413/413; 380-entry pack and clean install/init; six benchmark samples `ok`; version/syntax/diff/Cortex 8/8 | R14 baseline mitigated; R15 remains for WO-030 | Accepted locally; WO-027 Ready |
