+++ docs/agent-control/risk-register.md
@@ -19,6 +19,6 @@ the manager; the Status column carries the current truth, not the original fear.
 | R11 | Product migration | Removing MCP defaults before CLI retrieval has parity can strand existing users and weaken agent context quality | `cortex init --bootstrap` or existing Codex/Claude workflows may stop working, and agents may fall back to raw file reads | WO-016 through WO-019 implemented locally: CLI retrieval parity exists, MCP client registration is opt-in, `.context/mcp` remains the compatibility runtime path, and docs/package/release guidance now says CLI-first with MCP compatibility. Future package rename, `cortex mcp` removal, `.context/mcp` removal, or registry support removal requires a breaking-release or parallel-package plan. | Control Manager | Mitigated locally |
 | R12 | Semantic quality | Token compression, truncation, or retrieval-reranking changes can make Cortex faster or more diverse while losing important semantic signal | Cortex may look faster in benchmarks but return worse context to agents on real tasks | WO-020 now keeps compression behind explicit `CORTEX_EMBED_TEXT_PROFILE=compact-files`, compacts only large file-level records, preserves chunk text, and exposes profile/cache/manifest state. Cortex/bat quick gate and Angular capped-control gate had 0 lost expected hits. Cortex query pack was later calibrated to remove broken expected paths; accepted retrieval ranking improved Cortex to 29/35 and Angular to 22/42. Follow-up path/aspect/inflection experiments regressed and were rejected. The damaged Cortex/Angular benchmark workspaces were rebootstraped and sanity-checked back to 29/35 and 22/42 on 2026-06-25; rerun compact-files against the calibrated Cortex gate and run fastjson2 before promoting the profile beyond experimental opt-in. | Embedding and Validation | Partially mitigated |
 | R13 | Enterprise security | Remote skill identifiers and persisted paths could escape managed roots; license/cache/secret identity was weak; privileged commands trusted project runtime/state; per-user daemon services could cross project boundaries | Arbitrary root/user filesystem mutation, credential exposure or cross-org artifact reuse, and missing or misattributed governance | WO-024 adds strict ownership/containment, credential-bound cache/artifacts/events, TLS-only bearer egress, stdin-only atomic `0600` onboarding, package-owned privileged code with target-user project I/O, verified daemon control, and separated project/host services. Existing users explicitly re-enroll; legacy unmarked skills require reviewed backup/reconciliation. | Security and Privacy | Mitigated locally in 2.4.1; security and release reviews have no blocker/major findings |
-| R14 | Refactor integration | Moving CLI and ingest responsibilities can silently change command streams, exit codes, JSON envelopes, trusted Enterprise resolution, deterministic output, fallback behavior, or peak memory | Existing automation may break, Enterprise trust could regress, or large repositories could produce different context or fail | WO-026 accepted the executable 2.4.1 contract in `wo-026-characterization-baseline.md`. WO-027 preserved the CLI/Enterprise contract. WO-028 established one packaged canonical ingest implementation, preserved all four output digests plus sequential/parallel and worker-failure byte identity, passed focused/full/package/Cortex gates, and closed five-role review. WO-029 must keep worker, trace, ordering, retention, and repeated-memory gates green during orchestration extraction. | Control Manager + Validation | CLI and canonical/pure ingest extraction mitigated; worker/pipeline portion remains open through WO-029 |
+| R14 | Refactor integration | Moving CLI and ingest responsibilities can silently change command streams, exit codes, JSON envelopes, trusted Enterprise resolution, deterministic output, fallback behavior, or peak memory | Existing automation may break, Enterprise trust could regress, or large repositories could produce different context or fail | WO-026 accepted the executable 2.4.1 contract in `wo-026-characterization-baseline.md`. WO-027 preserved CLI/Enterprise behavior. WO-028 established the canonical ingest implementation. WO-029 extracted parser/worker/stage composition, preserved all digests, byte equivalence, worker/trace/retention contracts, passed focused/full/package gates, kept repeated median RSS within 5%, and fixed the review-discovered verbose-window state regression before five-role closure. | Control Manager + Validation | Mitigated locally through WO-029; WO-030/WO-031 retain integration gates for cleanup and release |
 | R15 | Scaffold ownership | Forced upgrades overlay new generated files but cannot distinguish obsolete Cortex-managed files from unknown user files | Stale runtime sources can break bootstrap, while broad cleanup could delete user configuration or secrets | WO-030 introduces a versioned managed-file manifest, contained path resolution, symlink/traversal denial, explicit preservation rules, and clean-upgrade fixtures; no cleanup ships before security and integration review | CLI and Scaffold + Security and Privacy | Open |
 | R16 | Ingest filesystem containment | Explicit absolute/parent/symlinked source paths can escape the project root, while cache/DB outputs and predictable staged JSONL paths do not reject symlink destinations | Local files outside the intended project can be copied into `.context`, or a hostile local filesystem layout can redirect ingest writes; ingest itself adds no network transmission | Preserve the accepted 2.4.1 behavior through WO-029. Create a separate security work order to define source-root containment, symlink policy, safe output opening/replacement, and failure cleanup with migration/negative tests before changing behavior. | Security and Privacy + Parsers and Ingest | Accepted pre-existing risk; unowned by WO-028/WO-029 |
